refactor(commands): create unified command entry point with auto-detection
Add unified bundle.ts and start.ts that merge the rspack and webpack
implementations, using detectBundler() and dynamic imports to select
the correct engine at runtime.

Add commands/index.ts exporting the unified command definitions and
a createBoundCommands() helper for deprecated entry points.

Add commands/index.js and commands/index.d.ts as the new primary
CJS entry point. Update commands/rspack.js and commands/webpack.js
to be deprecation wrappers that delegate via createBoundCommands().

Add './commands' export to package.json (alongside existing wildcard).

Delete old bundler-specific files:
- rspack/bundle.ts, rspack/start.ts, rspack/index.ts
- webpack/bundle.ts, webpack/start.ts, webpack/index.ts
