+++ packages/repack/commands/index.js
@@ -1,0 +1,2 @@
+const commands = require('../dist/commands/index.js');
+module.exports = commands.default;
