+++ packages/repack/commands/webpack.js
@@ -1,2 +1,5 @@
-const commands = require('../dist/commands/webpack/index.js');
-module.exports = commands.default;
+console.warn(
+  '[Re.Pack] Importing "@callstack/repack/commands/webpack" is deprecated. Use "@callstack/repack/commands" instead.'
+);
+const { createBoundCommands } = require('../dist/commands/index.js');
+module.exports = createBoundCommands('webpack');
