+++ packages/repack/src/commands/index.ts
@@ -1,6 +1,7 @@
-import { bundleCommandOptions, startCommandOptions } from '../options.js';
 import { bundle } from './bundle.js';
+import { bundleCommandOptions, startCommandOptions } from './options.js';
 import { start } from './start.js';
+import type { Bundler } from './types.js';
 
 const commands = [
   {
@@ -30,3 +31,16 @@ const commands = [
 ] as const;
 
 export default commands;
+
+/**
+ * Creates command definitions with a forced bundler engine.
+ * Used by deprecated entry points (`commands/rspack`, `commands/webpack`)
+ * to maintain backwards compatibility.
+ */
+export function createBoundCommands(bundler: Bundler) {
+  return commands.map((cmd) => ({
+    ...cmd,
+    func: (_: string[], cliConfig: any, args: any) =>
+      cmd.func(_, cliConfig, args, bundler),
+  }));
+}
